% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cpesr.R
\name{theme_cpesr}
\alias{theme_cpesr}
\title{theme_cpesr}
\usage{
theme_cpesr(base_family = "Raleway", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
ggplot theme for the CPESR work, based on theme_hc from ggthemes.
}
