% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cpesr.R
\name{theme_cpesr_setup}
\alias{theme_cpesr_setup}
\title{theme_cpesr_setup}
\usage{
theme_cpesr_setup(
  authors = NA,
  camille = TRUE,
  url = "cpesr.fr",
  licence = "GPL-3.0",
  base_family = "Raleway"
)
}
\arguments{
\item{authors}{a set of author names}

\item{camille}{add Camille Noûs as authors (default: TRUE)}

\item{url}{a referene url (default: cpesr.fr)}

\item{licence}{the licence (default: GPL-3.0)}

\item{base_family}{the font base family (default: Raleway)}

\item{...}{others arguments passed to theme_hc}
}
\value{

}
\description{
GGploth theme for the CPESR, adding a caption
}
\examples{
theme_cpesr_setup(authors=c("Author One","Author Two"),
                  camille=FALSE,
                  licence="CC BY-SA",
                  url="https://github.com/cpesr/cpesrthemes" )
}
